requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
Flask-Migrate==4.0.7
flask-login==0.6.3

